+++ src/main/java/ru/yandex/practicum/filmorate/dal/storage/film/InMemoryFilmStorage.java
@@ -18,11 +18,9 @@ import java.util.Map;
 public class InMemoryFilmStorage implements FilmStorage {
 
     @Getter
-    private final Map<Long, Film> films = new HashMap<>();
+    private final Map<Integer, Film> films = new HashMap<>();
 
     private int id = 1;
-
-
 
     public int getNextId() {
         return id++;
